requirements-dev.txt
# Development-only dependencies
# Install with: pip install -r requirements-dev.txt

# Include production requirements
-r requirements.txt

# Development & Debugging Tools
django-debug-toolbar>=4.2.0
django-extensions>=3.2.0

# Testing
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0
factory-boy>=3.3.0

# Code Quality
black>=23.0.0
flake8>=6.1.0
isort>=5.12.0
pylint>=3.0.0
pylint-django>=2.5.0

# Type Checking
mypy>=1.5.0
django-stubs>=4.2.0
